refactor: replace models with DTOs, update package structure, and streamline TeamService logic for better clarity and organization
