+++ brawl-core/src/main/java/brawl/example/project_brawl_api_sheets/integration_sheets/brawl_sheets/service/TeamService.java
@@ -1,33 +1,32 @@
 package brawl.example.project_brawl_api_sheets.integration_sheets.brawl_sheets.service;
 
-import brawl.example.project_brawl_api_sheets.integration_sheets.brawl_sheets.entity.PlayerTagData;
-import brawl.example.project_brawl_api_sheets.integration_sheets.brawl_sheets.model.TeamMODEL;
-import brawl.example.project_brawl_api_sheets.integration_sheets.brawl_sheets.model.TeamPlayerTag;
+import brawl.example.project_brawl_api_sheets.integration_sheets.brawl_sheets.entity.PlayerTagRepository;
+import brawl.example.project_brawl_api_sheets.integration_sheets.brawl_sheets.dto.TeamWithPlayersRelationDTO;
+import brawl.example.project_brawl_api_sheets.integration_sheets.brawl_sheets.util.intern_structure.TeamPlayerTag;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class TeamService {
     @Autowired
-    private PlayerTagData tagData;
-    //Aqui vamos pegar os dados do mongo vamos armazenar no team model e retornar a merda do team modeL(Chamdo de DTO)
+    private PlayerTagRepository tagData;
 
 
-    public List<TeamMODEL> getPlayersTagsANDnameTEAM() {
-        List<TeamPlayerTag> dados = tagData.findAllBy();
-        List<TeamMODEL> listTeamModel = new ArrayList<>();
+    public List<TeamWithPlayersRelationDTO> getPlayersTagsAndNameTeam() {
 
-        for (TeamPlayerTag time : dados) {
-            TeamMODEL model = new TeamMODEL();
-            model.setPlayersTags(time.getTags());
-            model.setTeamName(time.getTeamName());
+        List<TeamPlayerTag> dadosDoBanco = tagData.findAllBy();
 
-            listTeamModel.add(model);
-        }
 
-        return listTeamModel;
+        return dadosDoBanco.stream()
+                .map(timeDoBanco -> {
+                    TeamWithPlayersRelationDTO model = new TeamWithPlayersRelationDTO();
+                    model.setTeamName(timeDoBanco.getTeamName());
+                    model.setPlayersTags(timeDoBanco.getTags());
+                    return model;
+                })
+                .collect(Collectors.toList());
     }
 }
