+++ brawl-core/src/main/java/brawl/example/project_brawl_api_sheets/integration_sheets/brawl_sheets/util/intern_structure/TeamPlayerTag.java
@@ -1,4 +1,4 @@
-package brawl.example.project_brawl_api_sheets.integration_sheets.brawl_sheets.model;
+package brawl.example.project_brawl_api_sheets.integration_sheets.brawl_sheets.util.intern_structure;
 
 import java.util.List;
